3.x

+++ tests/Pager/PagePagerTest.php
@@ -14,7 +14,7 @@ final class PagePagerTest extends PagerTestCase
 {
     protected function createPager(array $results, int $page, int $limit): Pager
     {
-        $offset = (int) (($page - 1) * $limit);
+        $offset = ($page - 1) * $limit;
         $slice = \array_values(\array_slice($results, $offset, $limit));
 
         return new PagePager(new ArrayPage($slice, $offset, $limit, \count($results)));
